+++ gaiaxpy/cholesky/cholesky.py
@@ -6,11 +6,11 @@ from gaiaxpy.input_reader.input_reader import InputReader
 def __get_dot_product(L_inv):
     return dot(L_inv.T, L_inv)
 
+
 def __get_inv_cholesky_decomp_lower(xp_errors, xp_correlation_matrix):
     L = cholesky(xp_correlation_matrix, lower=True)
     # Invert lower triangular matrix.
-    ncoeffs = len(L)
-    L_inv = solve_triangular(L, identity(ncoeffs), lower=True)
+    L_inv = solve_triangular(L, identity(len(L)), lower=True)
     # Matrix of inverse errors.
     E_inv = diag(1.0 / xp_errors)
     return dot(L_inv, E_inv)
