+++ tests/test_cholesky/test_cholesky.py
@@ -3,10 +3,10 @@ import numpy as np
 import numpy.testing as npt
 import pandas as pd
 from os.path import join
-from gaiaxpy.cholesky.cholesky import get_chi2, get_inverse_covariance_matrix
 from gaiaxpy.core.generic_functions import array_to_symmetric_matrix
 from tests.files import files_path
 
+from gaiaxpy import get_chi2, get_inverse_covariance_matrix
 
 f = join(files_path, 'cholesky', '18Sco-XP-spectra.csv')
 # Load solution (only BP solution is available)
